injector/requirements.txt
typesense
python-dotenv
requests
